src/requirements.txt
celery==5.0.5
Django==3.1.5
django-cors-headers==3.7.0
django-environ==0.4.5
django-extensions==3.1.0
django-filter==2.4.0
django-fsm==2.7.1
djangorestframework==3.12.2
django-silk==4.1.0
future==0.18.2
gunicorn==20.0.4
ipython==7.19.0
packaging==20.8
psycopg2-binary==2.8.6
pyjwt==2.0.1
requests==2.25.1
sentry-sdk==0.19.5
drf-yasg==1.20.0
